+++ harry_potter/src/pages/mainPage/mainPage.tsx
@@ -1,13 +1,13 @@
 import { useContext } from 'react';
-import { ErrorModal } from '../../components/ErrorModal/errorModal';
-import { Footer } from '../../components/Footer/footer';
-import { Header } from '../../components/Header/header';
-import { SearchFieldComponent } from '../../components/Search/searchField';
-import { Spinner } from '../../components/Spinner/spinner';
-import { DataAppContext } from '../../context/dataAppContext';
-import { CardList } from '../../components/Results/cardList';
-import { Pangination } from '../../components/Pangination/pangination';
 import { useNavigate } from 'react-router';
+import { ErrorModal } from '$/components/ErrorModal';
+import { Footer } from '$/components/Footer';
+import { Header } from '$/components/Header';
+import { SearchFieldComponent } from '$/components/Search';
+import { Spinner } from '$/components/Spinner';
+import { DataAppContext } from '../../context/dataAppContext';
+import { CardList } from '$/components/Results';
+import { Pangination } from '$/components/Pangination';
 
 export function MainPage() {
   const { state, updateDetailesOpened, updateErrorThrow, updateShowModal } =
